Add drawIco16x16, fix updateInterval overflow, optimize mult. operations (#236)
* Fix updateInterval overflow when frame rate lower than 8
* Optimize operations mult. 2, 4, 8
* Add drawIco16x16 and setPixelColor functions
* return back clearPixel function for backward compatibility
* optimize clearPixel division

+++ src/OLEDDisplayUi.cpp
@@ -226,8 +226,7 @@ OLEDDisplayUiState* OLEDDisplayUi::getUiState(){
   return &this->state;
 }
 
-
-int8_t OLEDDisplayUi::update(){
+int16_t OLEDDisplayUi::update(){
 #ifdef ARDUINO
   unsigned long frameStart = millis();
 #elif __MBED__
@@ -237,7 +236,7 @@ int8_t OLEDDisplayUi::update(){
 #else
 #error "Unkown operating system"
 #endif
-  int8_t timeBudget = this->updateInterval - (frameStart - this->state.lastUpdate);
+  int16_t timeBudget = this->updateInterval - (frameStart - this->state.lastUpdate);
   if ( timeBudget <= 0) {
     // Implement frame skipping to ensure time budget is keept
     if (this->autoTransition && this->state.lastUpdate != 0) this->state.ticksSinceLastStateSwitch += ceil((double)-timeBudget / (double)this->updateInterval);
